EAMxx: moved field impl namespace to a details header
Keep field_impl.hpp smaller and easier to parse
